Back/requirements.txt
fastapi
uvicorn[standard]
python-multipart
bcrypt
python-jose[cryptography]
sqlalchemy
pydantic